refactored the chart context out into separate file; reverted CustomNode.tsx and created the new FlowChartAdvanced.tsx
