+++ stories/index.tsx
@@ -4,37 +4,38 @@ import { ConfigSnapToGridDemo } from './ConfigSnapToGrid'
 import { ConfigValidateLinkDemo } from './ConfigValidateLink'
 import { CustomCanvasOuterDemo } from './CustomCanvasOuter'
 import { CustomLinkDemo } from './CustomLink'
-import { Demo } from './CustomNode'
+import { CustomNodeDemo } from './CustomNode'
 import { CustomNodeInnerDemo } from './CustomNodeInner'
 import { CustomPortDemo } from './CustomPort'
 import { DragAndDropSidebar } from './DragAndDropSidebar'
 import { ExternalReactState } from './ExternalReactState'
+import { FlowChartAdvanced } from './FlowChartAdvanced'
 import { InternalReactState } from './InternalReactState'
 import { ReadonlyMode } from './ReadonlyMode'
 import { SelectedSidebar } from './SelectedSidebar'
 import { SmartRouting } from './SmartRouting'
 import { StressTestDemo } from './StressTest'
 
-// storiesOf('State', module)
-//   .add('Internal React State', InternalReactState)
-//   .add('External React State', () => <ExternalReactState />)
+storiesOf('State', module)
+  .add('Internal React State', InternalReactState)
+  .add('External React State', () => <ExternalReactState />)
 
 storiesOf('Custom Components', module)
-  // .add('Node Inner', () => <CustomNodeInnerDemo />)
-  // .add('Node -- Custom', () => <CustomNodeDemo />)
-  // .add('Port', CustomPortDemo)
-  // .add('Canvas Outer', CustomCanvasOuterDemo)
-  // .add('Canvas Link', () => <CustomLinkDemo />)
-  .add('CUSTOM', () => <Demo />)
+  .add('Node Inner', () => <CustomNodeInnerDemo />)
+  .add('Node', () => <CustomNodeDemo />)
+  .add('Port', CustomPortDemo)
+  .add('Canvas Outer', CustomCanvasOuterDemo)
+  .add('Canvas Link', () => <CustomLinkDemo />)
+  .add('Flow Chart Advanced', () => <FlowChartAdvanced />)
 
-// storiesOf('Stress Testing', module).add('default', StressTestDemo)
+storiesOf('Stress Testing', module).add('default', StressTestDemo)
 
-// storiesOf('Sidebar', module)
-//   .add('Drag and Drop', DragAndDropSidebar)
-//   .add('Selected Sidebar', () => <SelectedSidebar />)
+storiesOf('Sidebar', module)
+  .add('Drag and Drop', DragAndDropSidebar)
+  .add('Selected Sidebar', () => <SelectedSidebar />)
 
-// storiesOf('Other Config', module)
-//   .add('Snap To Grid', ConfigSnapToGridDemo)
-//   .add('Link validation function', ConfigValidateLinkDemo)
-//   .add('Read only mode', ReadonlyMode)
-//   .add('Smart link routing', SmartRouting)
+storiesOf('Other Config', module)
+  .add('Snap To Grid', ConfigSnapToGridDemo)
+  .add('Link validation function', ConfigValidateLinkDemo)
+  .add('Read only mode', ReadonlyMode)
+  .add('Smart link routing', SmartRouting)
